mention compatibility for old Julia versions
[skip ci]

For old Julia versions without artifacts, adding ImageMagick as a hard dependency can be a little bit tricky and unstable. Thus in the "best" practice in that time is to avoid the build process as possible as we can, and requesting users to install them manually.

ImageIO requires at least Julia 1.3.